每个脚本执行的进程为一个全新的从主进程fork的进程，子进程新引入的三方库及时通知主进程，避免后续重复引入

+++ script_execute_engine/main.py
@@ -12,6 +12,9 @@ import traceback
 import uuid
 import os
 
+original_lib_paths = sys.path.copy()
+original_lib_paths.remove(os.getcwd())
+
 sys.path.append("../scripts")
 #
 
@@ -19,26 +22,49 @@ sys.path.append("../scripts")
 app = Flask(__name__)
 
 
-def run_script(payload_dict):
+def load_module(module_name):
+    before_load_modules = set(sys.modules.keys())
+    run_module = importlib.import_module(module_name)
+    after_load_modules = set(sys.modules.keys())
+
+    # 把新引入的三方库通知主进程
+    new_import_modules = []
+    for module_key in after_load_modules - before_load_modules:
+        module = sys.modules.get(module_key)
+        if not module:
+            continue
+        try:
+            for original_lib_path in original_lib_paths:
+                if original_lib_path in module.__file__:
+                    new_import_modules.append(module_key)
+        except Exception as e:
+            new_import_modules.append(module_key)
+
+    return run_module, new_import_modules
+
+
+def run_script(payload_dict, conn):
     entrance_func_name = payload_dict.get("entrance_func")
     code = payload_dict.get("code")
 
-    code_file_name=f"c0dE_{str(uuid.uuid1())}"
+    code_file_name = f"c0dE_{str(uuid.uuid1())}"
 
     with open(f"../scripts/{code_file_name}.py", "w") as fd:
         fd.write(code)
     try:
-      module = importlib.import_module(code_file_name)
+        module, new_import_modules = load_module(code_file_name)
 
-      execute_func = getattr(module, entrance_func_name, None)
-      execute_result = execute_func()
-      return execute_result
+        execute_func = getattr(module, entrance_func_name, None)
+        execute_result = execute_func()
+
+        conn.send({"result": execute_result, "new_modules": new_import_modules})
     except Exception as e:
-      raise Exception("代码执行出错！")
+        conn.send({"result": str(e), "new_modules": []})
     finally:
-      os.remove(f"../scripts/{code_file_name}.py")
-    
-    
+        if os.path.exists(f"../scripts/{code_file_name}.py"):
+            os.remove(f"../scripts/{code_file_name}.py")
+        if os.path.exists(f"../scripts/__pycache__/{code_file_name}.cpython-37.pyc"):
+            os.remove(f"../scripts/__pycache__/{code_file_name}.cpython-37.pyc")
 
 
 # 编写路由，构建url与函数的映射关系（将函数与url绑定）
@@ -46,39 +72,33 @@ def run_script(payload_dict):
 def users():
     payload_dict = json.loads(request.data.decode("utf-8"))
     try:
-      p = pool.apply_async(run_script, (payload_dict,))
-      execute_result = p.get()
-      return jsonify({"code": 0, "message": "success", "result": execute_result})
+        parent_conn, child_conn = multiprocessing.Pipe()
+        p = multiprocessing.Process(target=run_script, args=(payload_dict, child_conn,))
+        p.start()
+        result = parent_conn.recv()
+        execute_result = result.get("result")
+        new_modules = result.get("new_modules")
+        for module_name in new_modules:
+            try:
+                m = __import__(module_name)
+                globals()[module_name] = m
+            except Exception as e:
+                pass
+        p.terminate()
+        p.kill()
+        p.join()
+
+        return jsonify({"code": 0, "message": "success", "result": execute_result})
     except Exception as e:
-      return jsonify({"code": 1, "message": "fail", "result": traceback.format_exc()})
-    
+        return jsonify({"code": 1, "message": "fail", "result": traceback.format_exc()})
 
 
 if __name__ == '__main__':
     # 默认方式启动
     # app.run()
     # 解决jsonify中文乱码问题
-    pool = multiprocessing.Pool(processes=4)
+    multiprocessing.set_start_method("fork")
 
     app.config['JSON_AS_ASCII'] = False
     # 以调试模式启动,host=0.0.0.0 ,则可以使用127.0.0.1、localhost以及本机ip来访问
     app.run(host="0.0.0.0", port=8899, debug=False)
-
-'''
-import requests
-import json
-
-url = "127.0.0.1:8899/execute"
-
-payload = json.dumps({
-  "entrance_func": "test_main",
-  "code": "def test_main():\n    return \"adsada###3212313251dfasasd\""
-})
-headers = {
-  'Content-Type': 'application/json'
-}
-
-response = requests.request("POST", url, headers=headers, data=payload)
-
-print(response.text)
-'''
